move: AutoCraftItemInputSignalReceiver -> InventoryMoveItemListener
