+++ src/main/java/io/github/sakaki_aruka/customcrafter/internal/listener/InventoryMoveItemListener.kt
@@ -1,0 +1,52 @@
+package io.github.sakaki_aruka.customcrafter.internal.listener
+
+import io.github.sakaki_aruka.customcrafter.CustomCrafterAPI
+import io.github.sakaki_aruka.customcrafter.internal.InternalAPI
+import io.github.sakaki_aruka.customcrafter.internal.autocrafting.CBlock
+import org.bukkit.Location
+import org.bukkit.Material
+import org.bukkit.World
+import org.bukkit.block.Block
+import org.bukkit.block.Crafter
+import org.bukkit.event.EventHandler
+import org.bukkit.event.Listener
+import org.bukkit.event.inventory.InventoryMoveItemEvent
+import org.bukkit.inventory.ItemStack
+
+object InventoryMoveItemListener: Listener {
+    @EventHandler
+    fun InventoryMoveItemEvent.onMove() {
+        if (!CustomCrafterAPI.getUseAutoCraftingFeature()) return
+        else if (this.isCancelled) return
+        else if (this.item.isEmpty) return
+        else if (this.destination.location == null) return
+        else if (this.destination.location!!.block.type != Material.CRAFTER) return
+
+        val crafter: Crafter = this.destination.location!!.block.state as? Crafter ?: return
+        if (!CBlock.hasEssentialKeys(crafter)) return
+        else if (!baseBlockCheck(crafter.block)) return
+
+        val cBlock: CBlock = CBlock.of(crafter) ?: return
+
+        if (!cBlock.addItems(this.item)) {
+            crafter.block.world.dropItem(cBlock.getDropLocation(), this.item)
+        }
+        this.item = ItemStack.empty()
+    }
+
+    private fun baseBlockCheck(crafter: Block): Boolean {
+        val crafterLoc: Location = crafter.location
+        val crafterWorld: World = crafter.world
+        val underCenter = Location(crafterWorld, crafterLoc.x, crafterLoc.y - 1, crafterLoc.z)
+        if (crafterWorld.getBlockAt(underCenter).type != Material.AIR) return false
+        val half: Int = InternalAPI.AUTO_CRAFTING_BASE_BLOCK_SIDE / 2
+        for (dz in (-half..half)) {
+            for (dx in (-half..half)) {
+                if (dx == 0 && dz == 0) continue
+                val loc = Location(crafterWorld, underCenter.x + dx, underCenter.y, underCenter.z + dz)
+                if (crafterWorld.getBlockAt(loc).type != CustomCrafterAPI.getAutoCraftingBaseBlock()) return false
+            }
+        }
+        return true
+    }
+}
